Resolved JS, Dockerfile, and WebSocket issues
JavaScript was unreachable from Go application. Root cause seems to have
been that Dockerfile was misconfigured. When deploying the application,
working directory is initially set to /usr/src/app, where app is then
built into a binary and executed. Auxillary files were then deleted.
This was a problem because the auxillary files included the HTML file
and the JavaScript file, both of which we still need during runtime.

To resolve this, I reconfigured Dockerfile to copy the Aux files to a
new folder, and make that new folder our working directory instead. This
circumvents the automated deletion that happens at the end of fly.io
build. Need to deep dive into Dockerfile further and see why this
behaviour exists.